# agent-verse-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentVerse",
    "version": "0.1.0"
  },
  "paths": {
...
    "/connectors": {
      "get": {
        "tags": [
          "connectors"
        ],
        "summary": "List Connectors",
        "operationId": "list_connectors_connectors_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Connectors Connectors Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "connectors"
        ],
        "summary": "Register Connector",
        "operationId": "register_connector_connectors_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterConnectorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Register Connector Connectors Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{server_id}": {
      "put": {
        "tags": [
          "connectors"
        ],
        "summary": "Update Connector",
        "operationId": "update_connector_connectors__server_id__put",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterConnectorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Connector Connectors  Server Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "connectors"
        ],
        "summary": "Unregister Connector",
        "operationId": "unregister_connector_connectors__server_id__delete",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{server_id}/test": {
      "post": {
        "tags": [
          "connectors"
        ],
        "summary": "Test Connector",
        "description": "Test connectivity to a registered MCP server by calling its /health endpoint.",
        "operationId": "test_connector_connectors__server_id__test_post",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Test Connector Connectors  Server Id  Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/oauth/start": {
      "get": {
        "tags": [
          "connectors"
        ],
        "summary": "Oauth Start",
        "description": "Start OAuth PKCE flow \u2014 returns the authorization URL with PKCE challenge.",
        "operationId": "oauth_start_connectors_oauth_start_get",
        "parameters": [
          {
            "name": "server_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Oauth Start Connectors Oauth Start Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/oauth/callback": {
      "get": {
        "tags": [
          "connectors"
        ],
        "summary": "Oauth Callback",
        "description": "OAuth callback \u2014 exchange authorization code for access tokens via PKCE.",
        "operationId": "oauth_callback_connectors_oauth_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "State"
            }
          },
          {
            "name": "server_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Server Id"
            }
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "http://localhost:8000/connectors/oauth/callback",
              "title": "Redirect Uri"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Oauth Callback Connectors Oauth Callback Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegisterConnectorRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "auth_type": {
            "type": "string",
            "title": "Auth Type"
          },
          "auth_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Auth Config",
            "default": {}
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "url",
          "auth_type"
        ],
        "title": "RegisterConnectorRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}